Move code related with findNodesThatFitPod() from generic_scheduler.go to scheduler.go
