+++ pkg/scheduler/generic_scheduler.go
@@ -21,19 +21,13 @@ import (
 	"fmt"
 	"math/rand"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	"k8s.io/klog/v2"
 
 	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/util/sets"
-	"k8s.io/apiserver/pkg/util/feature"
 	extenderv1 "k8s.io/kube-scheduler/extender/v1"
-	"k8s.io/kubernetes/pkg/features"
 	"k8s.io/kubernetes/pkg/scheduler/framework"
-	"k8s.io/kubernetes/pkg/scheduler/framework/parallelize"
-	"k8s.io/kubernetes/pkg/scheduler/framework/runtime"
 	internalcache "k8s.io/kubernetes/pkg/scheduler/internal/cache"
 	"k8s.io/kubernetes/pkg/scheduler/metrics"
 	utiltrace "k8s.io/utils/trace"
@@ -164,230 +158,6 @@ func (g *genericScheduler) selectHost(nodeScoreList framework.NodeScoreList) (st
 		}
 	}
 	return selected, nil
-}
-
-// numFeasibleNodesToFind returns the number of feasible nodes that once found, the scheduler stops
-// its search for more feasible nodes.
-func (g *genericScheduler) numFeasibleNodesToFind(numAllNodes int32) (numNodes int32) {
-	if numAllNodes < minFeasibleNodesToFind || g.percentageOfNodesToScore >= 100 {
-		return numAllNodes
-	}
-
-	adaptivePercentage := g.percentageOfNodesToScore
-	if adaptivePercentage <= 0 {
-		basePercentageOfNodesToScore := int32(50)
-		adaptivePercentage = basePercentageOfNodesToScore - numAllNodes/125
-		if adaptivePercentage < minFeasibleNodesPercentageToFind {
-			adaptivePercentage = minFeasibleNodesPercentageToFind
-		}
-	}
-
-	numNodes = numAllNodes * adaptivePercentage / 100
-	if numNodes < minFeasibleNodesToFind {
-		return minFeasibleNodesToFind
-	}
-
-	return numNodes
-}
-
-func (g *genericScheduler) evaluateNominatedNode(ctx context.Context, extenders []framework.Extender, pod *v1.Pod, fwk framework.Framework, state *framework.CycleState, diagnosis framework.Diagnosis) ([]*v1.Node, error) {
-	nnn := pod.Status.NominatedNodeName
-	nodeInfo, err := g.nodeInfoSnapshot.Get(nnn)
-	if err != nil {
-		return nil, err
-	}
-	node := []*framework.NodeInfo{nodeInfo}
-	feasibleNodes, err := g.findNodesThatPassFilters(ctx, fwk, state, pod, diagnosis, node)
-	if err != nil {
-		return nil, err
-	}
-
-	feasibleNodes, err = findNodesThatPassExtenders(extenders, pod, feasibleNodes, diagnosis.NodeToStatusMap)
-	if err != nil {
-		return nil, err
-	}
-
-	return feasibleNodes, nil
-}
-
-// Filters the nodes to find the ones that fit the pod based on the framework
-// filter plugins and filter extenders.
-func (g *genericScheduler) findNodesThatFitPod(ctx context.Context, extenders []framework.Extender, fwk framework.Framework, state *framework.CycleState, pod *v1.Pod) ([]*v1.Node, framework.Diagnosis, error) {
-	diagnosis := framework.Diagnosis{
-		NodeToStatusMap:      make(framework.NodeToStatusMap),
-		UnschedulablePlugins: sets.NewString(),
-	}
-
-	// Run "prefilter" plugins.
-	s := fwk.RunPreFilterPlugins(ctx, state, pod)
-	allNodes, err := g.nodeInfoSnapshot.NodeInfos().List()
-	if err != nil {
-		return nil, diagnosis, err
-	}
-	if !s.IsSuccess() {
-		if !s.IsUnschedulable() {
-			return nil, diagnosis, s.AsError()
-		}
-		// All nodes will have the same status. Some non trivial refactoring is
-		// needed to avoid this copy.
-		for _, n := range allNodes {
-			diagnosis.NodeToStatusMap[n.Node().Name] = s
-		}
-		// Status satisfying IsUnschedulable() gets injected into diagnosis.UnschedulablePlugins.
-		diagnosis.UnschedulablePlugins.Insert(s.FailedPlugin())
-		return nil, diagnosis, nil
-	}
-
-	// "NominatedNodeName" can potentially be set in a previous scheduling cycle as a result of preemption.
-	// This node is likely the only candidate that will fit the pod, and hence we try it first before iterating over all nodes.
-	if len(pod.Status.NominatedNodeName) > 0 && feature.DefaultFeatureGate.Enabled(features.PreferNominatedNode) {
-		feasibleNodes, err := g.evaluateNominatedNode(ctx, extenders, pod, fwk, state, diagnosis)
-		if err != nil {
-			klog.ErrorS(err, "Evaluation failed on nominated node", "pod", klog.KObj(pod), "node", pod.Status.NominatedNodeName)
-		}
-		// Nominated node passes all the filters, scheduler is good to assign this node to the pod.
-		if len(feasibleNodes) != 0 {
-			return feasibleNodes, diagnosis, nil
-		}
-	}
-	feasibleNodes, err := g.findNodesThatPassFilters(ctx, fwk, state, pod, diagnosis, allNodes)
-	if err != nil {
-		return nil, diagnosis, err
-	}
-
-	feasibleNodes, err = findNodesThatPassExtenders(extenders, pod, feasibleNodes, diagnosis.NodeToStatusMap)
-	if err != nil {
-		return nil, diagnosis, err
-	}
-	return feasibleNodes, diagnosis, nil
-}
-
-// findNodesThatPassFilters finds the nodes that fit the filter plugins.
-func (g *genericScheduler) findNodesThatPassFilters(
-	ctx context.Context,
-	fwk framework.Framework,
-	state *framework.CycleState,
-	pod *v1.Pod,
-	diagnosis framework.Diagnosis,
-	nodes []*framework.NodeInfo) ([]*v1.Node, error) {
-	numNodesToFind := g.numFeasibleNodesToFind(int32(len(nodes)))
-
-	// Create feasible list with enough space to avoid growing it
-	// and allow assigning.
-	feasibleNodes := make([]*v1.Node, numNodesToFind)
-
-	if !fwk.HasFilterPlugins() {
-		length := len(nodes)
-		for i := range feasibleNodes {
-			feasibleNodes[i] = nodes[(g.nextStartNodeIndex+i)%length].Node()
-		}
-		g.nextStartNodeIndex = (g.nextStartNodeIndex + len(feasibleNodes)) % length
-		return feasibleNodes, nil
-	}
-
-	errCh := parallelize.NewErrorChannel()
-	var statusesLock sync.Mutex
-	var feasibleNodesLen int32
-	ctx, cancel := context.WithCancel(ctx)
-	checkNode := func(i int) {
-		// We check the nodes starting from where we left off in the previous scheduling cycle,
-		// this is to make sure all nodes have the same chance of being examined across pods.
-		nodeInfo := nodes[(g.nextStartNodeIndex+i)%len(nodes)]
-		status := fwk.RunFilterPluginsWithNominatedPods(ctx, state, pod, nodeInfo)
-		if status.Code() == framework.Error {
-			errCh.SendErrorWithCancel(status.AsError(), cancel)
-			return
-		}
-		if status.IsSuccess() {
-			length := atomic.AddInt32(&feasibleNodesLen, 1)
-			if length > numNodesToFind {
-				cancel()
-				atomic.AddInt32(&feasibleNodesLen, -1)
-			} else {
-				feasibleNodes[length-1] = nodeInfo.Node()
-			}
-		} else {
-			statusesLock.Lock()
-			diagnosis.NodeToStatusMap[nodeInfo.Node().Name] = status
-			diagnosis.UnschedulablePlugins.Insert(status.FailedPlugin())
-			statusesLock.Unlock()
-		}
-	}
-
-	beginCheckNode := time.Now()
-	statusCode := framework.Success
-	defer func() {
-		// We record Filter extension point latency here instead of in framework.go because framework.RunFilterPlugins
-		// function is called for each node, whereas we want to have an overall latency for all nodes per scheduling cycle.
-		// Note that this latency also includes latency for `addNominatedPods`, which calls framework.RunPreFilterAddPod.
-		metrics.FrameworkExtensionPointDuration.WithLabelValues(runtime.Filter, statusCode.String(), fwk.ProfileName()).Observe(metrics.SinceInSeconds(beginCheckNode))
-	}()
-
-	// Stops searching for more nodes once the configured number of feasible nodes
-	// are found.
-	fwk.Parallelizer().Until(ctx, len(nodes), checkNode)
-	processedNodes := int(feasibleNodesLen) + len(diagnosis.NodeToStatusMap)
-	g.nextStartNodeIndex = (g.nextStartNodeIndex + processedNodes) % len(nodes)
-
-	feasibleNodes = feasibleNodes[:feasibleNodesLen]
-	if err := errCh.ReceiveError(); err != nil {
-		statusCode = framework.Error
-		return nil, err
-	}
-	return feasibleNodes, nil
-}
-
-func findNodesThatPassExtenders(extenders []framework.Extender, pod *v1.Pod, feasibleNodes []*v1.Node, statuses framework.NodeToStatusMap) ([]*v1.Node, error) {
-	// Extenders are called sequentially.
-	// Nodes in original feasibleNodes can be excluded in one extender, and pass on to the next
-	// extender in a decreasing manner.
-	for _, extender := range extenders {
-		if len(feasibleNodes) == 0 {
-			break
-		}
-		if !extender.IsInterested(pod) {
-			continue
-		}
-
-		// Status of failed nodes in failedAndUnresolvableMap will be added or overwritten in <statuses>,
-		// so that the scheduler framework can respect the UnschedulableAndUnresolvable status for
-		// particular nodes, and this may eventually improve preemption efficiency.
-		// Note: users are recommended to configure the extenders that may return UnschedulableAndUnresolvable
-		// status ahead of others.
-		feasibleList, failedMap, failedAndUnresolvableMap, err := extender.Filter(pod, feasibleNodes)
-		if err != nil {
-			if extender.IsIgnorable() {
-				klog.InfoS("Skipping extender as it returned error and has ignorable flag set", "extender", extender, "err", err)
-				continue
-			}
-			return nil, err
-		}
-
-		for failedNodeName, failedMsg := range failedAndUnresolvableMap {
-			var aggregatedReasons []string
-			if _, found := statuses[failedNodeName]; found {
-				aggregatedReasons = statuses[failedNodeName].Reasons()
-			}
-			aggregatedReasons = append(aggregatedReasons, failedMsg)
-			statuses[failedNodeName] = framework.NewStatus(framework.UnschedulableAndUnresolvable, aggregatedReasons...)
-		}
-
-		for failedNodeName, failedMsg := range failedMap {
-			if _, found := failedAndUnresolvableMap[failedNodeName]; found {
-				// failedAndUnresolvableMap takes precedence over failedMap
-				// note that this only happens if the extender returns the node in both maps
-				continue
-			}
-			if _, found := statuses[failedNodeName]; !found {
-				statuses[failedNodeName] = framework.NewStatus(framework.Unschedulable, failedMsg)
-			} else {
-				statuses[failedNodeName].AppendReason(failedMsg)
-			}
-		}
-
-		feasibleNodes = feasibleList
-	}
-	return feasibleNodes, nil
 }
 
 // prioritizeNodes prioritizes the nodes by running the score plugins,
